backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
requests==2.31.0
flake8==7.0.0
pytest==8.0.0
httpx==0.26.0
